remove jwlibrary.core replace eXtensionSharp

+++ JWLibrary.NUnit.Test/LiteDbFlexerTest.cs
@@ -1,43 +1,77 @@
-﻿using JLiteDBFlex;
-using JWLibrary.Core;
+﻿using eXtensionSharp;
+using JLiteDBFlex;
 using NUnit.Framework;
-using Service.Contract;
 
 namespace JWLibrary.NUnit.Test {
     public class LiteDbFlexerTest {
-        public CompanyDto companyDto = new CompanyDto();
-        
+        public CompanyDto companyDto = new();
+
         [SetUp]
         public void Setup() {
-            companyDto = new CompanyDto() {
+            companyDto = new CompanyDto {
                 Name = "red worker",
-                Ceo = new UserDto() {
+                Ceo = new UserDto {
                     Name = "kim",
-                    Age = 40,
+                    Age = 40
                 },
-                Cfo = new UserDto() {
+                Cfo = new UserDto {
                     Name = "hwang",
-                    Age = 50,
+                    Age = 50
                 },
-                Cto = new UserDto() {
+                Cto = new UserDto {
                     Name = "hong",
                     Age = 40
                 }
             };
         }
 
+        [Test]
+        public void get_companyinfo_test() {
+            var exists = LiteDbFlexerManager.Instance.Create<CompanyDto>()
+                .LiteCollection
+                .FindOne(m => m.Name == "red worker");
+
+            Assert.NotNull(exists);
+        }
+
+        [Test]
+        public void insert_companyinfo_test() {
+            var inserted = LiteDbFlexerManager.Instance.Create<CompanyDto>()
+                .LiteCollection
+                .Insert(companyDto);
+            Assert.Greater(inserted.AsInt32, 0);
+        }
+
+        [Test]
+        public void update_companyinfo_test() {
+            var exists = LiteDbFlexerManager.Instance.Create<CompanyDto>()
+                .LiteCollection.FindOne(m => m.Name == "red worker");
+
+            Assert.NotNull(exists);
+
+            var temp = exists.Cto;
+            exists.Cto = exists.Cfo;
+            exists.Cfo = temp;
+
+            var updated = LiteDbFlexerManager.Instance.Create<CompanyDto>()
+                .LiteCollection.Update(exists);
+
+            Assert.IsTrue(updated);
+        }
+
         #region [userdto test]
+
         [Test]
         public void insert_test() {
             var exists = LiteDbFlexerManager.Instance.Create<CompanyDto>().LiteCollection
                 .FindOne(m => m.Name == "red worker");
-            
+
             Assert.NotNull(exists);
-            
-            var ceoinserted= LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Insert(exists.Ceo);
+
+            var ceoinserted = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Insert(exists.Ceo);
             var cfoinserted = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Insert(exists.Cfo);
             var ctoinserted = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Insert(exists.Cto);
-            
+
             Assert.Greater(ceoinserted.AsInt32, 0);
             Assert.Greater(cfoinserted.AsInt32, 0);
             Assert.Greater(ctoinserted.AsInt32, 0);
@@ -52,24 +86,22 @@ namespace JWLibrary.NUnit.Test {
         [Test]
         public void get_delete_test() {
             var exists = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.FindOne(m => m.Name == "kim");
-            if (exists.jIsNotNull()) {
-                LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Delete(exists.Name);
-            }
+            if (exists.xIsNotNull()) LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Delete(exists.Name);
 
             exists = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.FindOne(m => m.Name == "kim");
-            
+
             Assert.IsNull(exists);
         }
 
         [Test]
         public void get_update_test() {
             var exists = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.FindOne(m => m.Name == "hwang");
-            if (exists.jIsNotNull()) {
+            if (exists.xIsNotNull()) {
                 exists.Age = 22;
-                var result= LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Update(exists);
+                var result = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.Update(exists);
                 Assert.IsTrue(result);
             }
-            
+
             exists = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection.FindOne(m => m.Name == "hwang");
             Assert.AreEqual(22, exists.Age);
         }
@@ -85,62 +117,26 @@ namespace JWLibrary.NUnit.Test {
                 var updated = LiteDbFlexerManager.Instance.Create<UserDto>()
                     .LiteCollection
                     .Update(exists);
-                
+
                 Assert.IsTrue(updated);
 
                 exists = LiteDbFlexerManager.Instance.Create<UserDto>()
                     .LiteCollection
                     .FindOne(m => m.Name == "test");
-                
+
                 Assert.AreEqual(40, exists.Age);
 
                 var removed = LiteDbFlexerManager.Instance.Create<UserDto>().LiteCollection
                     .Delete(exists.Name);
-                
+
                 Assert.IsTrue(removed);
             }
         }
-        
 
         #endregion
-       
-        [Test]
-        public void get_companyinfo_test() {
-            var exists = LiteDbFlexerManager.Instance.Create<CompanyDto>()
-                .LiteCollection
-                .FindOne(m => m.Name == "red worker");
-	
-            Assert.NotNull(exists);            
-        }
-
-        [Test]
-        public void insert_companyinfo_test() {
-            var inserted = LiteDbFlexerManager.Instance.Create<CompanyDto>()
-                .LiteCollection
-                .Insert(companyDto);
-            Assert.Greater(inserted.AsInt32, 0);            
-        }
-
-        [Test]
-        public void update_companyinfo_test() {
-            var exists = LiteDbFlexerManager.Instance.Create<CompanyDto>()
-                .LiteCollection.FindOne(m => m.Name == "red worker");
-            
-            Assert.NotNull(exists);
-
-            var temp = exists.Cto;
-            exists.Cto = exists.Cfo;
-            exists.Cfo = temp;
-
-            var updated = LiteDbFlexerManager.Instance.Create<CompanyDto>()
-                .LiteCollection.Update(exists);
-            
-            Assert.IsTrue(updated);
-        }
-        
     }
 
-    [LiteDbTable("cominfo.db", "user", new [] {"Name"})]
+    [LiteDbTable("cominfo.db", "user", new[] {"Name"})]
     public class UserDto {
         //auto-increment
         //id가 명시적으로 없을 경우 _oid를 생성함. 해당 hash값이 고유값이 됨. Key와는 다른 의미임.
@@ -157,7 +153,4 @@ namespace JWLibrary.NUnit.Test {
         public UserDto Cfo { get; set; }
         public UserDto Cto { get; set; }
     }
-    
-    
-    
 }
